[REFACTOR] : 디자인, 레디 확인 로직, 코스 추가하는 로직, 점수 반영 로직 수정, RoomRecord API 관련 코드 추가

+++ YogaYo/domain/src/main/java/com/d104/domain/model/PeerUser.kt
@@ -4,6 +4,6 @@ data class PeerUser(
     val id: String,
     val nickName: String,
     val isReady: Boolean = false,
-    val totalScore: Float = 0.0f,
+    val totalScore: Int = 0,
     val roundScore: Float = 0.0f
 )
